requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
itsdangerous==2.2.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
typing_extensions==4.12.1
